#380 プレイヤー設定画面での再生速度の設定UIをListViewに変更

+++ NicoPlayerHohoema/Views/Converters/ToUserFriendlyNumber.cs
@@ -19,7 +19,7 @@ namespace NicoPlayerHohoema.Views.Converters
 			if (value is uint)		{ return ((uint)value).ToString("N0"); }
 			if (value is ulong)		{ return ((ulong)value).ToString("N0"); }
 			if (value is float)		{ return ((float)value).ToString("N2"); }
-			if (value is double)	{ return ((int)value).ToString("N2"); }
+			if (value is double)	{ return ((double)value).ToString("N2"); }
 
 			return value;
 		}
